\name{BrCa_1_AR}
\alias{BrCa_1_AR}
\docType{data}
\title{
Breast cancer 1-Attributable Risk
}
\description{
   1-Attributable Risk
}
\usage{data("BrCa_1_AR")}
\format{
  A data frame with 2 observations on the following 5 variables.
  \describe{
    \item{\code{Wh.Gail}}{White}
    \item{\code{AA.CARE}}{African-American}
    \item{\code{Hi.Gail}}{Hispanic-American}
    \item{\code{NA.Gail}}{Native American and unknown race}
    \item{\code{Asian.AABCS}}{Asian-American}
  }
}

