\name{BrCa_beta}
\alias{BrCa_beta}
\docType{data}
\title{
Breast cancer beta
}
\description{
The logistic regression coefficients derived from the Gail model.
}
\usage{data("BrCa_beta")}
\format{
  A data frame with 6 observations on the following 5 variables.
  \describe{
    \item{\code{Wh.Gail}}{White, Gail model}
    \item{\code{AA.CARE}}{African-American, Care model}
    \item{\code{Hi.Gail}}{Hispanic-American, Gail model}
    \item{\code{NA.Gail}}{Native American and unknown race, Gail model}
    \item{\code{Asian.AABCS}}{Asian-American, AABCS model}
  }
}

