\name{BrCa_lambda2}
\alias{BrCa_lambda2}
\docType{data}
\title{
Breast cancer competing mortality
}
\description{
Breast cancer competing mortality for different races and age groups 
from 20 to 90 by 5 years.
}
\usage{data("BrCa_lambda2")}
\format{
  A data frame with 14 age groups on the following 12 variables.
  \describe{
    \item{\code{Wh.1983_87}}{White SEER 1983:1987}
    \item{\code{AA.1994_98}}{African-American SEER 1994:1998}
    \item{\code{Hi.1990_96}}{Hispanic-American 1990:1996}
    \item{\code{NA.1983_87}}{Native American and unknown race 1983:1987}
    \item{\code{Wo.1995_03}}{White SEER 1995:2003}
    \item{\code{Ch.1998_02}}{Chinese-American SEER 18 1998:2002}
    \item{\code{Ja.1998_02}}{Japanese-American SEER 18 1998:2002}
    \item{\code{Fi.1998_02}}{Filipino-American SEER 18 1998:2002}
    \item{\code{Hw.1998_02}}{Hawaiian SEER 18 1998:2002}
    \item{\code{oP.1998_02}}{Other Pacific Islander SEER 18 1998:2002}
    \item{\code{oA.1998_02}}{Other Asian SEER 1998:2002}
    \item{\code{Wh_Avg.1992_96}}{Average White SEER 1992:1996}
  }
}