% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVarsBCRA.R
\name{getDatosVariable}
\alias{getDatosVariable}
\title{getDatosVariable}
\usage{
getDatosVariable(idVariable, desde, hasta)
}
\arguments{
\item{idVariable}{id de la variable a obtener}

\item{desde}{fecha desde la cual se quiere obtener los datos}

\item{hasta}{fecha hasta la cual se quiere obtener los datos}
}
\value{
un tibble con los datos de la variable
}
\description{
Obtiene los datos de una variable del BCRA
}
