% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVarsBCRA.R
\name{getPrincipalesVariables}
\alias{getPrincipalesVariables}
\title{getPrincipalesVariables}
\usage{
getPrincipalesVariables()
}
\description{
Obtiene las principales variables del BCRA

 @return un tibble con las principales variables del BCRA
}
